fix: refactor to follow SOLID principle
